backend/requirements.txt
alembic
asgi_lifespan
argon2_cffi
asyncio
asyncpg
bcrypt
black
colorama
email-validator
pydantic_settings
fastapi
greenlet
httpx
isort
loguru
mypy
passlib
pathlib
pre-commit
python-dotenv
python-decouple
python-jose
python-slugify
pytest
pytest-asyncio
pytest-cov
pytest-xdist
python-multipart
SQLAlchemy==2.0.0b3
trio
uvicorn
